dashboard/requirements.txt
streamlit==1.31.0
cassandra-driver==3.29.0
pandas==2.2.1
plotly==5.20.0
